Move state definition to processing_threads.h
